Sort out the state of the sdk regen database and move args.js to a better location.
