+++ test/genengine.test.js
@@ -19,7 +19,7 @@
  */
 
 const genEngine = require('../src-electron/generator/generation-engine.js')
-const args = require('../src-electron/main-process/args.js')
+const args = require('../src-electron/util/args.js')
 const env = require('../src-electron/util/env.js')
 const dbApi = require('../src-electron/db/db-api.js')
 const fs = require('fs')
