+++ test/validation.test.js
@@ -22,7 +22,7 @@ const fs = require('fs')
 const dbApi = require('../src-electron/db/db-api.js')
 const zclLoader = require('../src-electron/zcl/zcl-loader.js')
 const Validation = require('../src-electron/validation/validation.js')
-const args = require('../src-electron/main-process/args.js')
+const args = require('../src-electron/util/args.js')
 const querySession = require('../src-electron/db/query-session.js')
 const queryConfig = require('../src-electron/db/query-config.js')
 const queryZcl = require('../src-electron/db/query-zcl.js')
